HOt fix of mean function

+++ stupid_math.c
@@ -1,5 +1,5 @@
 #include "stupid_math.h"
 
 float mean(int a, int b){
-	return (a+b)/2;
+	return (float)(a+b)/(float)2;
 }
